+++ common/src/main/java/com/niuktok/backend/common/entity/UserCommentLike.java
@@ -14,13 +14,13 @@ import lombok.ToString;
 public class UserCommentLike implements Serializable {
     @Id
     @GeneratedValue(generator = "JDBC")
-    private Integer id;
+    private Long id;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "comment_id")
-    private Integer commentId;
+    private Long commentId;
 
     @Column(name = "created_time")
     private Date createdTime;
